+++ src/Z38/SwissPayment/TransactionInformation/BankCreditTransfer.php
@@ -2,6 +2,7 @@
 
 namespace Z38\SwissPayment\TransactionInformation;
 
+use DOMDocument;
 use Z38\SwissPayment\BC;
 use Z38\SwissPayment\BIC;
 use Z38\SwissPayment\FinancialInstitutionInterface;
@@ -46,7 +47,7 @@ class BankCreditTransfer extends CreditTransfer
     /**
      * {@inheritdoc}
      */
-    public function asDom(\DOMDocument $doc, PaymentInformation $paymentInformation)
+    public function asDom(DOMDocument $doc, PaymentInformation $paymentInformation)
     {
         $root = $this->buildHeader($doc, $paymentInformation);
 
@@ -57,9 +58,7 @@ class BankCreditTransfer extends CreditTransfer
         $root->appendChild($this->buildCreditor($doc));
 
         $creditorAccount = $doc->createElement('CdtrAcct');
-        $creditorAccountId = $doc->createElement('Id');
-        $creditorAccountId->appendChild($doc->createElement('IBAN', $this->creditorIBAN->normalize()));
-        $creditorAccount->appendChild($creditorAccountId);
+        $creditorAccount->appendChild($this->creditorIBAN->asDom($doc));
         $root->appendChild($creditorAccount);
 
         if ($this->hasRemittanceInformation()) {